Incomplete/Untrialed Power Five Card Function
Has now been successful in early trials - Sums how many fives have been played, and multiplies the value of the five to the next player. Forces the player to pick up if the current card is a five and if they have no other card playable in that situation.

Tests show that bugs are still present such as the unnecessary compiling of five_count_compilation